feat: create model containers from LabSDK
This change introduces support for creation of model custom containers from the LabSDK seamlessly.

In addition, it refactors the labsdk.raptor.types, and introduce support for sagemaker serverless
