+++ hack/check-headers-for-license.sh
@@ -20,8 +20,8 @@ set -o pipefail
 
 function listPkgDirs() {
 	go list -f '{{.Dir}}' -tags=ignore_autogenerated,e2e ./... | grep -v generated
-	find labsdk
-	find runtime
+	git ls-files labsdk
+	git ls-files runtime
 }
 
 #Lists all go files
