+++ labsdk/raptor/_internal/exporter/__init__.py
@@ -12,27 +12,5 @@
 #  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 #  See the License for the specific language governing permissions and
 #  limitations under the License.
-import os
-from typing import Optional
 
-
-class BaseModelFramework:
-    @staticmethod
-    def save(model, spec: 'ModelSpec'):
-        raise NotImplementedError
-
-    @staticmethod
-    def predict(model, spec: 'ModelSpec'):
-        raise NotImplementedError
-
-
-    @staticmethod
-    def _base_output_path():
-        return f'out/models'
-
-    @staticmethod
-    def _create_output_path(path: Optional[str] = None):
-        if path is None:
-            path = BaseModelFramework._base_output_path()
-        if not os.path.exists(path):
-            os.makedirs(path)
+from .model import *
