+++ labsdk/raptor/types/model.py
@@ -12,20 +12,35 @@
 #  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 #  See the License for the specific language governing permissions and
 #  limitations under the License.
+#
+#  Licensed under the Apache License, Version 2.0 (the "License");
+#  you may not use this file except in compliance with the License.
+#  You may obtain a copy of the License at
+#
+#      http://www.apache.org/licenses/LICENSE-2.0
+#
+#  Unless required by applicable law or agreed to in writing, software
+#  distributed under the License is distributed on an "AS IS" BASIS,
+#  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+#  See the License for the specific language governing permissions and
+#  limitations under the License.
+import os
+from datetime import timedelta
 from enum import Enum
-from typing import List, Callable
+from typing import List, Callable, Optional
 from warnings import warn
 
 import pandas as pd
 import yaml
 
-from .frameworks.huggingface import HuggingFaceFramework
-from .frameworks.sklearn import SklearnFramework
-from .frameworks.xgboost import XGBoostFramework
+from .common import RaptorSpec, EnumSpec
 from .yaml import RaptorDumper
+from .._internal import model_servers
+
+os.environ.setdefault('BENTOML_HOME', os.path.join(os.path.expanduser('~'), '.raptor', 'bentoml'))
 
 
-class ModelServer(Enum):
+class ModelServer(EnumSpec):
     SageMakerACK = 'sagemaker-ack'
     Seldon = 'seldon'
     KServe = 'kserve'
@@ -48,32 +63,31 @@ class ModelServer(Enum):
         return self.value
 
     @classmethod
-    def to_yaml(cls, dumper: yaml.dumper.Dumper, data: 'ModelServer'):
-        return dumper.represent_scalar('!ModelServer', data.value)
+    def to_yaml_dict(cls, data: 'ModelServer'):
+        return data.value
 
     @property
-    def config(self):
+    def config(self) -> Optional[model_servers.ModelServer]:
         if self == ModelServer.SageMakerACK:
-            return {
-                'region': '$AWS_REGION',
-                'executionRoleARN': '$AWS_EXECUTION_ROLE_ARN',
-            }
-        return {}
+            return model_servers.Sagemaker
 
-
-RaptorDumper.add_representer(ModelServer, ModelServer.to_yaml)
+        warn(f'ModelServer {self} is not supported yet')
+        return None
 
 
 class ModelFramework(Enum):
     HuggingFace = 'huggingface'
     Sklearn = 'sklearn'
     Pytorch = 'pytorch'
+    PytorchLightning = 'pytorch-lightning'
     Tensorflow = 'tensorflow'
     CatBoost = 'catboost'
     ONNX = 'onnx'
     LightGBM = 'lightgbm'
     XGBoost = 'xgboost'
     FastAI = 'fastai'
+    Keras = 'keras'
+    Picklable = 'picklable'
 
     @staticmethod
     def parse(m):
@@ -84,15 +98,6 @@ class ModelFramework(Enum):
                 if e.value == m:
                     return e
         raise Exception(f'Unknown ModelFramework {m}')
-
-    def save(self, model, spec):
-        if self == ModelFramework.HuggingFace:
-            return HuggingFaceFramework.save(model, spec)
-        elif self == ModelFramework.Sklearn:
-            return SklearnFramework.save(model, spec)
-        elif self == ModelFramework.XGBoost:
-            return XGBoostFramework.save(model, spec)
-        warn(f'ModelFramework {self} not Deployable by Raptor at the moment. Please export your model manually instead')
 
     @classmethod
     def to_yaml(cls, dumper: yaml.dumper.Dumper, data: 'ModelFramework'):
@@ -124,3 +129,26 @@ class TrainingContext:
         Get the features and labels for the model
         """
         return self._data_getter()
+
+
+class ModelSpec(RaptorSpec):
+    keys: [str] = None
+    freshness: Optional[timedelta] = None
+    staleness: timedelta = None
+    timeout: timedelta = None
+    features: [str] = None
+    label_features: [str] = None
+
+    key_feature: str = None
+
+    model_framework: ModelFramework = None
+    model_server: Optional[ModelServer] = None
+
+    training_function: Callable = None
+    # noinspection PyUnresolvedReferences
+    exporter: 'ModelExporter' = None
+
+    _model_image: str = None
+
+    def train(self):
+        raise NotImplementedError()
